Refactor sendLoginEmail() -> email-service module
Issue: Scottish-Natural-Heritage/Licensing#1962
